requirements.txt
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.11
faiss-cpu==1.12.0
python-dotenv==1.0.1
streamlit==1.37.1
tiktoken==0.7.0
pytest==8.3.2
